+++ src/main/kotlin/io/extremum/ground/client/storage/HeadersHolder.kt
@@ -1,0 +1,6 @@
+package io.extremum.ground.client.storage
+
+object HeadersHolder {
+
+    val headers = ThreadLocal<Map<String, String>?>()
+}